Proper fixed for not overriding type default values on showOnAttach.
Clean up Material SideNav structures (PUsh type) removed any api related to it from Base SideNav class.
